% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{Client}
\alias{Client}
\title{Client}
\description{
Client

Client
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{api_key}}{Arize API key}

\item{\code{organization_key}}{Arize organization key}

\item{\code{files_uri}}{Arize API URI}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Client$new()}}
\item \href{#method-post_file}{\code{Client$post_file()}}
\item \href{#method-log}{\code{Client$log()}}
\item \href{#method-clone}{\code{Client$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Init a new client instance
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$new(
  api_key = Sys.getenv("ARIZE_API_KEY"),
  organization_key = Sys.getenv("ARIZE_ORGANIZATION_KEY"),
  uri = "https://api.arize.com/v1"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{api_key}}{character, the Arize-AI api key}

\item{\code{organization_key}}{character, the Arize-AI organization key}

\item{\code{uri}}{character, the API URI}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
A check is made for environmental variables for the API and Organization keys
via Sys.getenv("ARIZE_API_KEY") and Sys.getenv("ARIZE_ORGANIZATION_KEY"). To set these
environmental variables, use, e.g., Sys.setenv("ARIZE_ORGANIZATION_KEY" = "YOUR_ARIZE_ORGANIZATION_KEY")
or create an \code{.Renviron} file and list these two keys.
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post_file"></a>}}
\if{latex}{\out{\hypertarget{method-post_file}{}}}
\subsection{Method \code{post_file()}}{
Post a serialized file
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$post_file(.path, .schema, .sync)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.path}}{path to the file to post}

\item{\code{.schema}}{the schema (see \code{?arize::create_schema})}

\item{\code{.sync}}{logical, whether to sync}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log"></a>}}
\if{latex}{\out{\hypertarget{method-log}{}}}
\subsection{Method \code{log()}}{
Log to Arize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$log(
  .data_frame,
  .model_id,
  .model_type,
  .environment,
  .schema,
  .model_version = NULL,
  .batch_id = NULL,
  .sync = FALSE,
  .validate = TRUE,
  .path = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{.data_frame}}{data.frame to log}

\item{\code{.model_id}}{character, id for the model}

\item{\code{.model_type}}{integer, \code{1} for binary, \code{2} for numeric, \code{3} for categorical, and \code{4} for score-categorical}

\item{\code{.environment}}{.environment integer, \code{1} for production, \code{2} for validation, \code{3} for training}

\item{\code{.schema}}{the schema (see \code{?arize::create_schema})}

\item{\code{.model_version}}{character, the model version, optional}

\item{\code{.batch_id}}{character, the batch id, optional}

\item{\code{.sync}}{logical, whether to sync}

\item{\code{.validate}}{logical, whether to run validation checks}

\item{\code{.path}}{character, path to use for serialization, optional}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Client$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
