% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{create_schema}
\alias{create_schema}
\title{Schema}
\usage{
create_schema(.list = NULL, ...)
}
\arguments{
\item{.list}{A named list of key-value pairs for schema fields}

\item{...}{Key-value pairs of schema fields}
}
\description{
Create a Schema list
}
\details{
Field options are:
\itemize{
\item prediction_id_column_name
\item feature_column_names*
\item timestamp_column_name
\item prediction_label_column_name
\item prediction_score_column_name
\item actual_label_column_name
\item actual_score_column_name
\item shap_values_column_names*
}

Fields denoted with a star are lists of column names
}
\examples{
create_schema(timestamp_column_name = "timestamp")
create_schema(.list = list(timestamp_column_name = "timestamp"))
create_schema(
  .list = list(timestamp_column_name = "timestamp"),
  prediction_id_column_name = "predict_id"
)
create_schema(
  feature_column_names = c(colnames(iris)[1:4], "Stamen.Count", "Anther.Length"),
  prediction_id_column_name = "Species"
)
}
