% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{environments}
\alias{environments}
\title{Arize AI environments}
\format{
An object of class \code{list} of length 3.
}
\usage{
environments
}
\description{
A named list with enivronments as names and integer values 1:3
used to set the environment when logging to Arize AI
}
\keyword{datasets}
