% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{model_types}
\alias{model_types}
\title{Arize AI model types}
\format{
An object of class \code{list} of length 4.
}
\usage{
model_types
}
\description{
A named list with model types as names and integer values 1:4
used to set the model type when logging to Arize AI
}
\keyword{datasets}
