% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{schema_fields}
\alias{schema_fields}
\title{Arize AI schema fields}
\format{
An object of class \code{list} of length 8.
}
\usage{
schema_fields
}
\description{
A list of fields that can be supplied to \code{create_schema}
}
\keyword{datasets}
