% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validate_params}
\alias{validate_params}
\title{validate_params}
\usage{
validate_params(
  .data_frame,
  .model_id,
  .model_type,
  .environment,
  .schema,
  .model_version,
  .batch_id
)
}
\arguments{
\item{.data_frame}{the data}

\item{.model_id}{character, the id of the model}

\item{.model_type}{The type of the model, see the \code{arize::model_types} data object}

\item{.environment}{integer, \code{1} for production, \code{2} for validation, \code{3} for training}

\item{.schema}{the schema (see \code{?arize::create_schema})}

\item{.model_version}{character, the model version}

\item{.batch_id}{character, the batch id}
}
\description{
Wrapper to run parameter validation functions
}
