% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validate_types}
\alias{validate_types}
\title{validate_types}
\usage{
validate_types(.model_type, .schema, .arrow_schema)
}
\arguments{
\item{.model_type}{The type of the model, see the \code{arize::model_types} data object}

\item{.schema}{the schema (see \code{?arize::create_schema})}

\item{.arrow_schema}{named_list, with column names as names and column types as values}
}
\description{
Wrapper to run type validation functions
}
