% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{validate_values}
\alias{validate_values}
\title{validate_values}
\usage{
validate_values(.data_frame, .schema)
}
\arguments{
\item{.data_frame}{the data}

\item{.schema}{the schema (see \code{?arize::create_schema})}
}
\description{
Wrapper to run value validation functions
}
