% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{Validator}
\alias{Validator}
\title{Validator}
\description{
R6 class to validate data frame and schema
}
\examples{
I <- iris
I$Species_predicted_label <- I$Species
I$Species_predicted_score <- 1:150
I$Species_actual_label <- I$Species
I$Species_actual_score <- 150:1
I$Timestamp <- Sys.Date()

S <- create_schema(
  prediction_id_column_name = "Species",
  feature_column_names = colnames(I)[1:4],
  timestamp_column_name = "Timestamp",
  prediction_label_column_name = "Species_predicted_label",
  prediction_score_column_name = "Species_predicted_score",
  actual_label_column_name = "Species_actual_label",
  actual_score_column_name = "Species_actual_score"
)

V <- Validator$new()
V$validateParameters(
  .data_frame = I,
  .model_id = "1",
  .model_type = model_types$CATEGORICAL,
  .environment = environments$TRAINING,
  .schema = S,
  .model_version = "1",
  .batch_id = "1"
)
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-validateParameters}{\code{Validator$validateParameters()}}
\item \href{#method-validateTypes}{\code{Validator$validateTypes()}}
\item \href{#method-validateValues}{\code{Validator$validateValues()}}
\item \href{#method-clone}{\code{Validator$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validateParameters"></a>}}
\if{latex}{\out{\hypertarget{method-validateParameters}{}}}
\subsection{Method \code{validateParameters()}}{
Run parameter validation functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validator$validateParameters(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named arguments sent to \code{validate_params} (see \code{?validate_params})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validateTypes"></a>}}
\if{latex}{\out{\hypertarget{method-validateTypes}{}}}
\subsection{Method \code{validateTypes()}}{
Run parameter validation functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validator$validateTypes(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named arguments sent to \code{validate_types} (see \code{?validate_types})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validateValues"></a>}}
\if{latex}{\out{\hypertarget{method-validateValues}{}}}
\subsection{Method \code{validateValues()}}{
Run parameter validation functions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validator$validateValues(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{Named arguments sent to \code{validate_values} see( \code{?validate_values})}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Validator$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
